Share common class fixtures
