+++ tests/Integration/Database/EloquentMorphToEagerLoadTest.php
@@ -6,6 +6,7 @@ use Illuminate\Contracts\Database\Eloquent\CastsAttributes;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Tests\Database\Fixtures\Models\Comment;
 use Illuminate\Tests\Integration\Database\DatabaseTestCase;
 
 class EloquentMorphToEagerLoadTest extends DatabaseTestCase
@@ -99,16 +100,6 @@ class Article extends Model
     protected $casts = ['slug' => ArticleSlug::class];
 
     protected $fillable = ['slug'];
-}
-
-class Comment extends Model
-{
-    public $timestamps = false;
-
-    public function commentable()
-    {
-        return $this->morphTo();
-    }
 }
 
 class Uuid
