+++ tests/Integration/Routing/RouteCachingTest.php
@@ -8,7 +8,7 @@ class RouteCachingTest extends TestCase
 {
     public function testWildcardCatchAllRoutes()
     {
-        $this->routes(__DIR__.'/Fixtures/wildcard_catch_all_routes.php');
+        $this->routes(dirname(__DIR__, 2).'/Routing/Fixtures/wildcard_catch_all_routes.php');
 
         $this->get('/foo')->assertSee('Regular route');
         $this->get('/bar')->assertSee('Wildcard route');
@@ -16,7 +16,7 @@ class RouteCachingTest extends TestCase
 
     public function testRedirectRoutes()
     {
-        $this->routes(__DIR__.'/Fixtures/redirect_routes.php');
+        $this->routes(dirname(__DIR__, 2).'/Routing/Fixtures/redirect_routes.php');
 
         $this->post('/foo/1')->assertRedirect('/foo/1/bar');
         $this->get('/foo/1/bar')->assertSee('Redirect response');
