+++ tests/Integration/Support/SupportMailTest.php
@@ -2,8 +2,8 @@
 
 namespace Illuminate\Tests\Integration\Support;
 
-use Illuminate\Mail\Mailable;
 use Illuminate\Support\Facades\Mail;
+use Illuminate\Tests\Mail\Fixtures\TestMail;
 use Orchestra\Testbench\TestCase;
 
 class SupportMailTest extends TestCase
@@ -38,18 +38,5 @@ class SupportMailTest extends TestCase
         Mail::to('hello@laravel.com')->send(new TestMail());
 
         Mail::assertSent(TestMail::class);
-    }
-}
-
-class TestMail extends Mailable
-{
-    /**
-     * Build the message.
-     *
-     * @return $this
-     */
-    public function build()
-    {
-        return $this->view('view');
     }
 }
