+++ tests/Mail/Fixtures/TestMail.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Illuminate\Tests\Mail\Fixtures;
+
+use Illuminate\Mail\Mailable;
+
+class TestMail extends Mailable
+{
+    /**
+     * Build the message.
+     *
+     * @return $this
+     */
+    public function build()
+    {
+        return $this->view('view');
+    }
+}
