+++ tests/Notifications/Fixtures/NotificationStub.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Illuminate\Tests\Notifications\Fixtures;
+
+use Illuminate\Notifications\Notification;
+
+class NotificationStub extends Notification
+{
+    public function via($notifiable)
+    {
+        return ['mail'];
+    }
+}
