+++ tests/Routing/Fixtures/Enums.php
@@ -14,11 +14,6 @@ enum CategoryBackedEnum: string
     case Fruits = 'fruits';
 }
 
-enum RouteNameEnum: string
-{
-    case UserIndex = 'users.index';
-}
-
 enum RouteDomainEnum: string
 {
     case DashboardDomain = 'dashboard.myapp.com';
